channels: Introduce new lastdata callback
The lastdata callback is called when a channel is
about to be closed. It gives the application developer
a final chance to process the remaining buffered data,
if it did not consume all bytes in a previous data
callback invocation.

The lastdata callback is called when an eof, exit-status,
exit-signal or close message arrives for the channel,
whatever comes first. The close message is the only
mandatory message to close a channel. All others are
optional. However, it does not make much sense if the
lastdata callback is called _after_ the eof, exit-status
or exit-signal callback. Therefore, it is called as
early as reasonable.

If a lastdata callback is set, the channel state is
always set to closed even if not all data was consumed
by the callback.

Signed-off-by: Tilo Eckert <tilo.eckert@flam.de>